+++ NEWS.md
@@ -4,6 +4,8 @@
   
   * add function blockBootSample for block bootstrapping with panel data
 
+  * add option in makeDist to force the values of the distribution function be between 0 and 1
+
 # BMisc 1.3.1
 
   * Update rhs.vars to fix bug related to formulas like y~x+I(x^2)